requirements.txt
torch>=2.1
torchvision>=0.16
numpy
tqdm
pyyaml
opencv-python
matplotlib
scipy
pycocotools
